Lazy-seq-based version. Removed core.aysnc.
Fixed a bug where catching a transactor timeout and retrying would not
actually retry - it would skip the tx and go on to the next one. This
turned out to be hard to fix with the channel-based main-loop, so I
switched to using a lazy-seq of transactions. This thing is
single-threaded anyway, so doing it this way makes sense.

In testing with a real database, I made it through a few 100K of
transactions, but then I got timeout exceptions about 10 times in a row,
followed by a "transactor unavailable". I'll have to figure out how to
handle that one - maybe just pause for a much longer time (a minute or
so?) to let the transactor GC, or index, or do whatever it needs to do.